Finished & Added Event documentation 10/05/2019 | 19w40a4 | 0.5.45

+++ TRAP/src/Graphics/Buffers/BufferLayout.h
@@ -48,7 +48,7 @@ namespace TRAP::Graphics
 	{
 		std::string Name;
 		ShaderDataType Type{ShaderDataType::None};
-		uint32_t Size{};
+		intptr_t Size{};
 		uint32_t Offset{};
 		bool Normalized{};
